Split out byte array decoders (#2318)
