+++ parquet/src/arrow/mod.rs
@@ -123,6 +123,7 @@ experimental!(mod array_reader);
 pub mod arrow_reader;
 pub mod arrow_writer;
 mod buffer;
+mod decoder;
 
 #[cfg(feature = "async")]
 pub mod async_reader;
